Eayunstack 1.0 minor fixes for iso building

+++ image/centos/module.mk
@@ -28,7 +28,7 @@ $(BUILD_DIR)/images/$(TARGET_CENTOS_IMG_ART_NAME):
 	mkdir -p $(BUILD_DIR)/image/centos
 	sudo sh -c "$${SANDBOX_UP}"
 	sudo yum -c $(SANDBOX)/etc/yum.conf --installroot=$(SANDBOX) -y --nogpgcheck install tar python-setuptools git python-imgcreate python-argparse PyYAML
-	sudo cp /etc/mtab $(SANDBOX)/etc/mtab
+	if ! [ -e /etc/mtab ]; then sudo cp /etc/mtab $(SANDBOX)/etc/mtab; fi
 	sudo mkdir -p $(SANDBOX)/run/shm
 	sudo cp $(SOURCE_DIR)/image/centos/build_centos_image.py $(SANDBOX)/build_centos_image.py
 	sudo cp $(SOURCE_DIR)/image/centos/centos.ks $(SANDBOX)/centos.ks
